# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_keel-hq_keel_pkg_auth.AuthResponse:
    properties:
      token:
        type: string
    type: object
  pkg_http.APIResponse:
    properties:
      status:
        type: string
    type: object
  pkg_http.ApproveRequest:
    properties:
      action:
        description: defaults to approve
        type: string
      id:
        type: string
      identifier:
        type: string
      voter:
        type: string
    type: object
  pkg_http.AuditLogsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/types.AuditLog'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  pkg_http.EmptyResponse:
    type: object
...
  pkg_http.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  pkg_http.ResourceApprovalsUpdateRequest:
    properties:
      identifier:
        type: string
      provider:
        type: string
      votesRequired:
        type: integer
    type: object
...
  types.Approval:
    properties:
      archived:
        description: Archived is set to true once approval is finally approved/rejected
        type: boolean
      createdAt:
        description: When this approval was created
        type: string
      currentVersion:
        type: string
      deadline:
        description: Deadline for this request
        type: string
      digest:
        description: |-
          Digest is used to verify that images are the ones that got the approvals.
          If digest doesn't match for the image, votes are reset.
        type: string
      event:
        allOf:
        - $ref: '#/definitions/types.Event'
        description: Event that triggered evaluation
      id:
        type: string
      identifier:
        description: |-
          Identifier is used to inform user about specific
          Helm release or k8s deployment
          ie: k8s <namespace>/<deployment name>
              helm: <namespace>/<release name>
        type: string
      message:
        type: string
      newVersion:
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/types.ProviderType'
        description: Provider name - Kubernetes/Helm
      rejected:
        description: |-
          Explicitly rejected approval
          can be set directly by user
          so even if deadline is not reached approval
          could be turned down
        type: boolean
      updatedAt:
        description: WHen this approval was updated
        type: string
      voters:
        allOf:
        - $ref: '#/definitions/types.JSONB'
        description: |-
          Voters is a list of voter
          IDs for audit
      votesReceived:
        type: integer
      votesRequired:
        description: |-
          Requirements for the update such as number of votes
          and deadline
        type: integer
    type: object
  types.AuditLog:
    properties:
      accountId:
        type: string
      action:
        description: create/delete/update
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      identifier:
        type: string
      message:
        type: string
      metadata:
        $ref: '#/definitions/types.JSONB'
      payload:
        description: can be used for bigger messages such as webhook payload
        type: string
      payloadType:
        type: string
      resourceKind:
        description: approval/deployment/daemonset/statefulset/etc...
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  types.Event:
    properties:
      createdAt:
        type: string
      repository:
        $ref: '#/definitions/types.Repository'
      triggerName:
        description: optional field to identify trigger
        type: string
    type: object
  types.JSONB:
    additionalProperties: true
    type: object
  types.ProviderType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - ProviderTypeUnknown
    - ProviderTypeKubernetes
    - ProviderTypeHelm
  types.Repository:
    properties:
      digest:
        description: optional digest field
        type: string
      host:
        type: string
      name:
        type: string
      tag:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API exposed by Keel. Admin routes are registered only when an
    authenticator is enabled. Provider webhooks require Basic or Bearer authorization
    only when authenticated webhooks are enabled; the registry webhook is always unauthenticated.
  title: Keel HTTP API
  version: "1.0"
paths:
...
  /v1/approvals:
    get:
      description: Lists active and archived approvals. This route exists only when
        the authenticator is enabled. Legacy store and serialization error paths write
        text with status 200.
      operationId: listApprovals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Approval'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List approvals
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Approves, rejects, deletes, or archives an approval. Delete returns
        JSON null; legacy delete/archive error paths write text with status 200. This
        route exists only when the authenticator is enabled.
      operationId: updateApproval
      parameters:
      - description: Approval action
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.ApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Approval'
        "400":
          description: Malformed request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Approval not found
          schema:
            type: string
        "500":
          description: Approval update failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update an approval
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Sets a Kubernetes resource's required vote count. This route exists
        only when the authenticator is enabled.
      operationId: setResourceApprovals
      parameters:
      - description: Approval requirement
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.ResourceApprovalsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.APIResponse'
        "400":
          description: Malformed or unsupported request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Set required approvals
      tags:
      - Admin
  /v1/audit:
    get:
      description: Lists audit entries. Invalid limit or offset values are silently
        treated as zero. This route exists only when the authenticator is enabled.
      operationId: listAuditLogs
      parameters:
      - description: Maximum entries
        in: query
        name: limit
        type: integer
      - description: Entry offset
        in: query
        name: offset
        type: integer
      - description: Comma-separated resource kinds
        in: query
        name: filter
        type: string
      - description: Account email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.AuditLogsResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "500":
          description: Store query failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List audit logs
      tags:
      - Admin
...
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates Basic credentials supplied as JSON. This route exists
        only when the authenticator is enabled.
      operationId: login
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer token
              type: string
          schema:
            $ref: '#/definitions/github_com_keel-hq_keel_pkg_auth.AuthResponse'
        "400":
          description: Malformed request
          schema:
            type: string
        "401":
          description: Incorrect username or password
          schema:
            type: string
      summary: Log in
      tags:
      - Auth
  /v1/auth/logout:
    get:
      description: Returns an empty JSON object. This route exists only when the authenticator
        is enabled.
      operationId: logoutViaGet
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Log out with GET
      tags:
      - Auth
    post:
      description: Returns an empty JSON object. This route exists only when the authenticator
        is enabled.
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Log out
      tags:
      - Auth
  /v1/auth/refresh:
    get:
      description: Generates a new bearer token. This route exists only when the authenticator
        is enabled.
      operationId: refreshAuth
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Refreshed bearer token
              type: string
          schema:
            $ref: '#/definitions/github_com_keel-hq_keel_pkg_auth.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Token generation failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Refresh authentication
      tags:
      - Auth
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"